+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/shopkeepers/ShopkeepersCommand.java
@@ -79,6 +79,7 @@ public class ShopkeepersCommand extends BaseCommand {
 		childCommands.register(new CommandRemoveAll(plugin, shopkeeperRegistry, confirmations));
 		childCommands.register(new CommandGive());
 		childCommands.register(new CommandGiveCurrency());
+		childCommands.register(new CommandSetCurrency());
 		childCommands.register(new CommandConvertItems());
 		childCommands.register(new CommandRemote());
 		childCommands.register(new CommandEdit());